chore: 死んでいるBlockItem.tsx（sidebar）を削除する
apps/product/src/components/shell/sidebar/BlockItem.tsxはproduction consumer
がゼロで、参照元はStorybook storyのみだった（knipはstory経由の参照があるため
dead code扱いしないが、grep実測で production 側の呼び出しはゼロと確認）。
BlockItem.tsx / BlockItem.stories.tsxを削除し、SidebarSection.stories.tsx内の
使用箇所（16箇所）はfile-local mockコンポーネントへ置き換えて見た目を維持した。

Closes #2164
